Split EditorTextView+Rendering.swift into focused files
The rendering file had grown to ~930 lines. Move self-contained groups into
sibling files (pure code movement, no behavior change), and document each:

- EditorTextView+ListRendering.swift  — list/checkbox/bullet markers + indent
- EditorTextView+TableSupport.swift   — TableRowTextBlock + row/cell parsing
- EditorTextView+MathRendering.swift  — now also holds colorMathSource +
  displayMathParagraphStyle (raw-LaTeX coloring belongs with math)

EditorTextView+Rendering.swift is now ~565 lines (the styling primitives,
the styleBlock dispatch switch, and in-place restyling) with an expanded
header explaining the architecture and where each piece now lives.

Symbols used across files were widened from private to internal.
